+++ pyxel/run.py
@@ -9,6 +9,7 @@
 import logging
 import sys
 import time
+import warnings
 from collections.abc import Sequence
 from pathlib import Path
 from typing import TYPE_CHECKING, Optional, Union
@@ -24,7 +25,7 @@ from pyxel.detectors import APD, CCD, CMOS, MKID, Detector
 from pyxel.exposure import Exposure
 from pyxel.observation import Observation, ObservationResult
 from pyxel.pipelines import DetectionPipeline, Processor
-from pyxel.util import create_model, download_examples
+from pyxel.util import create_model, deprecated, download_examples
 
 if TYPE_CHECKING:
     import xarray as xr
@@ -35,12 +36,19 @@ if TYPE_CHECKING:
 
 
 # TODO: This function will be deprecated (see #563)
+@deprecated(
+    "Model 'pyxel.models.optics.load_psf' is deprecated and will be removed in version"
+    " 2. Use model 'pyxel.models.photon_collection.stripe_pattern' instead."
+)
 def exposure_mode(
     exposure: "Exposure",
     detector: Detector,
     pipeline: "DetectionPipeline",
 ) -> "xr.Dataset":
     """Run an 'exposure' pipeline.
+
+    .. deprecated:: 1.8.0
+        `exposure_mode` will be removed in pyxel 2.0.0, it is replaced by `run_mode`.
 
     For more information, see :ref:`exposure_mode`.
 
@@ -83,6 +91,7 @@ def exposure_mode(
         signal        (readout_time, y, x) float64 3.159 3.03 2.971 ... 3.195 3.36
         pixel         (readout_time, y, x) float64 1.053e+03 1.01e+03 ... 1.12e+03
     """
+    warnings.warn("Use function 'pyxel.run_mode'", DeprecationWarning, stacklevel=1)
 
     logging.info("Mode: Exposure")
 
@@ -187,6 +196,9 @@ def observation_mode(
 ) -> "ObservationResult":
     """Run an 'observation' pipeline.
 
+    .. deprecated:: 1.8.0
+        `observation_mode` will be removed in pyxel 2.0.0, it is replaced by `run_mode`.
+
     For more information, see :ref:`observation_mode`.
 
     Parameters
@@ -219,6 +231,8 @@ def observation_mode(
     >>> result
     ObservationResult(...)
     """
+    warnings.warn("Use function 'pyxel.run_mode'", DeprecationWarning, stacklevel=1)
+
     logging.info("Mode: Observation")
 
     observation_outputs: ObservationOutputs = observation.outputs
@@ -246,6 +260,9 @@ def calibration_mode(
     compute_and_save: bool = True,
 ) -> tuple["xr.Dataset", pd.DataFrame, pd.DataFrame, Sequence]:
     """Run a 'calibration' pipeline.
+
+    .. deprecated:: 1.8.0
+        `calibration_mode` will be removed in pyxel 2.0.0, it is replaced by `run_mode`.
 
     For more information, see :ref:`calibration_mode`.
 
@@ -327,6 +344,8 @@ def calibration_mode(
     []
     """
     # Late import to speedup start-up time
+    warnings.warn("Use function 'pyxel.run_mode'", DeprecationWarning, stacklevel=1)
+
     from pyxel.calibration import CalibrationResult
 
     logging.info("Mode: Calibration")