+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Pyxel doesn't use SemVer anymore, since most minor releases have at least minor 
   (See [!787](https://gitlab.com/esa/pyxel/-/merge_requests/787)).
 * Improved method `Scene.to_xarray()`.
   (See [!799](https://gitlab.com/esa/pyxel/-/merge_requests/799)).
+* Added deprecated warnings to `pyxel.calibdation_mode()`, `pyxel.exposure_mode()`, `pyxel.observation_mode()`.
+  (See [!802](https://gitlab.com/esa/pyxel/-/merge_requests/802)).
 
 ### Documentation
 * Add documentation for `Detector.scene`.